+++ src/utils/utilities.ts
@@ -1,5 +1,6 @@
 import ACSParticipantDisplayName from "../core/messaging/ACSParticipantDisplayName";
 import MessageType from "@microsoft/omnichannel-ic3core/lib/model/MessageType";
+import { uuidv4 } from "..";
 
 export const isSystemMessage = (message: any): boolean => { // eslint-disable-line @typescript-eslint/no-explicit-any,  @typescript-eslint/explicit-module-boundary-types
     const {messageType, properties} = message;
@@ -26,4 +27,15 @@ export const isCustomerMessage = (message: any): boolean => { // eslint-disable-
 export const isClientIdNotFoundErrorMessage = (e: any): boolean => {
     return e?.response?.status === 401
         && e?.response?.headers?.message === "UserId not found";
+}
+
+export const isNotEmpty = (value: string | null) : boolean => {
+    return value !== null && value !== undefined && value.trim() !== '';
+}
+
+export const getRuntimeId = (externalRuntimeId : string | null ): string => {
+    if (externalRuntimeId !== null && isNotEmpty(externalRuntimeId)) {
+        return externalRuntimeId;
+    }
+    return uuidv4();
 }